requirements.txt
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-WTF==0.14.2
psycopg2==2.8.2
pg8000==1.13.1
pymongo==3.8.0
pymongo[gssapi]==0.7.0
pymongo[srv]==1.16.0
pymongo[tls]==3.8.0
pymongo[snappy]==0.5.4
pymongo[zstd]==3.8.0
dnspython==1.16.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.3
google-api-core==1.14.2
google-cloud-pubsub==0.45.0
grpc-google-iam-v1==0.12.3
google-api-python-client==1.7.10
pytz==2019.2
mixpanel==4.4.0
PyJWT==1.7.1
twilio==6.29.4
firebase-admin==2.17.0


# dev
#coverage==4.5.3
#flake8==3.7.7
#Flask-Testing==0.7.1
